+++ src/geometry/vec2d.h
@@ -3,19 +3,19 @@
 
 class Vec2d {
  public:
-  float x, y;
+  double x, y;
 
   Vec2d();
-  Vec2d(float x_, float y_);
+  Vec2d(double x_, double y_);
 
   Vec2d operator+(const Vec2d& other);
   Vec2d& operator+=(const Vec2d& other);
   Vec2d operator-(const Vec2d& other);
   Vec2d& operator-=(const Vec2d& other);
-  Vec2d operator*(const float& other);
+  Vec2d operator*(const double& other);
   Vec2d& operator*=(const Vec2d& other);
-  Vec2d& operator*=(const float& other);
-  Vec2d& operator/=(const float& other);
+  Vec2d& operator*=(const double& other);
+  Vec2d& operator/=(const double& other);
 };
 
 #endif  // !VEC2D_H
